+++ smushclient-qt/rust/src/ffi/client.rs
@@ -85,14 +85,14 @@ pub mod ffi {
         fn connect_to_host(self: &SmushClient, socket: Pin<&mut QAbstractSocket>);
         fn handle_connect(self: &SmushClient, socket: Pin<&mut QAbstractSocket>) -> QString;
         fn handle_disconnect(self: Pin<&mut SmushClient>);
-        fn world_option(self: &SmushClient, index: usize, option: &[u8]) -> i32;
+        fn world_option(self: &SmushClient, index: usize, option: &[u8]) -> i64;
         fn world_alpha_option(self: &SmushClient, index: usize, option: &[u8]) -> VariableView;
         fn world_variant_option(self: &SmushClient, index: usize, option: &[u8]) -> QVariant;
         fn set_world_option(
             self: Pin<&mut SmushClient>,
             index: usize,
             option: &[u8],
-            value: i32,
+            value: i64,
         ) -> ApiCode;
         fn set_world_alpha_option(
             self: Pin<&mut SmushClient>,
